pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sourcenet"
version = "0.1.0"
description = "Deep Learning for Moment Tensor Inversion"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
  {name = "", email = ""}
]

# Core dependencies required to run the model
# Using >= versions ensures compatibility with newer environments
dependencies = [
    "torch>=2.0.0",
    "numpy",
    "scipy",
    "pandas",
    "h5py",
    "obspy",           # For seismological signal processing
    "matplotlib",      # For plotting beachballs and curves
    "tqdm",            # For progress bars
    "hydra-core>=1.3", # For configuration management (CRITICAL for L5)
    "omegaconf",
    "scikit-learn",    # For train_test_split
]

# Optional dependencies for development, testing, and formatting
[project.optional-dependencies]
dev = [
    "pytest",          # Testing framework
    "pytest-cov",      # Coverage reporting
    "black",           # Code formatter
    "isort",           # Import sorter
    "flake8",          # Linter
    "mypy",            # Static type checker
    "jupyterlab",      # For notebooks
]

# --- Tool Configuration ---

[tool.setuptools.packages.find]
where = ["src"]
include = ["sourcenet*"]
namespaces = false

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
python_files = "test_*.py"
